% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_words.R
\name{filter_words}
\alias{filter_words}
\title{Replace word prefixes that occur in given words.}
\usage{
filter_words(.data, .words, .replacement, ...)
}
\arguments{
\item{.data}{A data.frame or tibble.}

\item{.words}{A list of words.}

\item{.replacement}{A character string used as replacement.}

\item{...}{Columns to act upon.}
}
\value{
The original data, filtered.
}
\description{
Finds words starting with a word in a given list, and
replaces each such word prefix with a given replacement.
}
\examples{
data <- data.frame(
  x = c("This is some text...", "...containing words."),
  y = c("This is more text...", "...containing something."),
  z = c("This is some more text...", "...containing  more words.")
)

data <- data \%>\%
  filter_words(
    c("some", "words"),
    "####",
    x, y
  )

#                      x                        y                         z
# 1 This is #### text... This is more text...     This is some more text...
# 2 ...containing ####.  ...containing ####thing. ...containing more words.
}
