% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profanityfilter-package.R
\docType{package}
\name{profanityfilter-package}
\alias{profanityfilter}
\alias{profanityfilter-package}
\title{profanityfilter: Filter Profanity From Text}
\description{
Fast profanity filtering using a trie data structure. Also exports
  a filter_words function that can be used in a dplyr pipeline.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MarkMc1089/profanityfilter}
  \item Report bugs at \url{https://github.com/MarkMc1089/profanityfilter/issues}
}

}
\keyword{internal}
