% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfanityFilter.R
\name{ProfanityFilter}
\alias{ProfanityFilter}
\title{R6 Class providing a profanity filter to censor text.}
\description{
Uses a trie data structure of profane words to filter profanity
from text.
}
\examples{

## ------------------------------------------------
## Method `ProfanityFilter$new`
## ------------------------------------------------

\dontrun{
  # use default profanity list
  pf <- ProfanityFilter$new()

  # use custom profanity list
  bad_words <- list("feck", "sod", "damn")
  pf <- ProfanityFilter$new(bad_words)
}

## ------------------------------------------------
## Method `ProfanityFilter$censor`
## ------------------------------------------------

\dontrun{
  pf$censor("Don't say feck or bugger", '+')
  # "Don't say ++++ or ++++++!"
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProfanityFilter$new()}}
\item \href{#method-censor}{\code{ProfanityFilter$censor()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ProfanityFilter} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfanityFilter$new(
  profane_words = rjson::fromJSON(file = system.file("extdata", "profane_words.json",
    package = "profanityfilter"))
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{profane_words}}{Unnamed list of words (by default it reads the included
\code{profane_words.json} into a list).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ProfanityFilter} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  # use default profanity list
  pf <- ProfanityFilter$new()

  # use custom profanity list
  bad_words <- list("feck", "sod", "damn")
  pf <- ProfanityFilter$new(bad_words)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-censor"></a>}}
\if{latex}{\out{\hypertarget{method-censor}{}}}
\subsection{Method \code{censor()}}{
Apply profanity filtering to a piece of text.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfanityFilter$censor(text, censor_char = "*")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{The text to filter.}

\item{\code{censor_char}}{Character used to replace all characters in censored
words (default '*').}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The original text, filtered of profanity.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  pf$censor("Don't say feck or bugger", '+')
  # "Don't say ++++ or ++++++!"
}
}
\if{html}{\out{</div>}}

}

}
}
