% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trie.R
\name{Trie}
\alias{Trie}
\title{R6 Class providing a trie data structure}
\description{
Trie is implemented as nested unnamed lists, with a special
character string flag denoting end of a word.
}
\examples{

## ------------------------------------------------
## Method `Trie$new`
## ------------------------------------------------

words <- list("first", "fist", "fjord")
t <- Trie$new(words)

## ------------------------------------------------
## Method `Trie$has_prefix`
## ------------------------------------------------

t$has_prefix("firstly")
# TRUE
t$has_prefix("firs")
# FALSE

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Trie$new()}}
\item \href{#method-has_prefix}{\code{Trie$has_prefix()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Trie} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trie$new(words)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{words}}{Unnamed list of words.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Trie} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{words <- list("first", "fist", "fjord")
t <- Trie$new(words)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_prefix"></a>}}
\if{latex}{\out{\hypertarget{method-has_prefix}{}}}
\subsection{Method \code{has_prefix()}}{
Check if a given word is present in the trie.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trie$has_prefix(word)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{word}}{String to check for presence in trie.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Either \code{TRUE} or \code{FALSE}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{t$has_prefix("firstly")
# TRUE
t$has_prefix("firs")
# FALSE

}
\if{html}{\out{</div>}}

}

}
}
